+++ frontend/pages/ManageControlsPage/OSSettings/cards/Certificates/Certificates.tsx
@@ -6,7 +6,7 @@ import { formatDistanceToNow } from "date-fns";
 import { AppContext } from "context/app";
 import PATHS from "router/paths";
 
-import UploadList from "pages/ManageControlsPage/components/UploadList";
+import UploadList from "components/UploadList";
 import UploadListHeading from "pages/ManageControlsPage/components/UploadListHeading";
 
 import GitOpsModeTooltipWrapper from "components/GitOpsModeTooltipWrapper";
