+++ frontend/pages/ManageControlsPage/SetupExperience/cards/BootstrapPackage/components/UploadedPackageView/UploadedPackageView.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { IBootstrapPackageMetadata } from "interfaces/mdm";
 
 import CustomLink from "components/CustomLink";
-import UploadList from "pages/ManageControlsPage/components/UploadList";
+import UploadList from "components/UploadList";
 
 import BootstrapPackageListItem from "../BootstrapPackageListItem";
 import BootstrapPackageTable from "../BootstrapPackageTable/BootstrapPackageTable";
